Displacement component bug fix thanks to Hanna Elston
Correct sloppy check for the existence of vertical velocities.
Previously, if they were not part of the input structure, then
a field of all zeros was created, but dcomp was set to [1 1 0]
so that they're not used. But, this didn't check for a previous
existence of dcomp. Now, dcomp's existence is checked, and if
not specified as an input argument, it's set to [1 1 1] to use
all 3 components, followed by modification of just the third
entry if the vertical velocities are empty.